Otimizando classes

+++ supermercado.bll/src/supermercado/bll/CategoriaEJB.java
@@ -18,12 +18,11 @@ public class CategoriaEJB implements ICategoriaEJB{
 
 		try {
 			categoriaDAO.insertOrUpdate(categoria);
+			return new Mensagem("Salvo com sucesso.", MensagemStatus.sucesso);
 		}catch(Exception ex) {
 			return new Mensagem("Ocorreu um erro inesperado: " 
 						+ ex.getMessage(),MensagemStatus.erro);
 		}
-		
-		return new Mensagem("Salvo com sucesso.", MensagemStatus.sucesso);
 	}
 
 	public Mensagem excluir(Short idCategoria) {
@@ -37,15 +36,17 @@ public class CategoriaEJB implements ICategoriaEJB{
 				throw new Exception("Categoria inexistente.");
 			}
 			
+			if(categoria.getProdutos().size() > 0) {
+				throw new Exception("Essa categoria possui produtos vinculados a ela.");
+			}
 			categoriaDAO.delete(categoria);
+			return new Mensagem("Categoria Excluída.",
+					MensagemStatus.sucesso);
 			
 		}catch(Exception ex) {
 			return new Mensagem("Não foi possível excluir: " 
 					+ ex.getMessage(), MensagemStatus.erro);
 		}
-		
-		return new Mensagem("Excluído com sucesso.",
-				MensagemStatus.sucesso);
 		
 	}
